% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_3wks.R
\name{air_3wks}
\alias{air_3wks}
\title{The AI rate in the last 3 weeks}
\usage{
air_3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
double  AI rate in the last 3 weeks(\%)
}
\description{
((the_sum_of_AI -  the_sum_of_the_unknown_pregnant_AI) /
(the_number_of_cycles -
the_number_of_the_unknown_pregnancy_cycles ))*100 \%
..in_the_last_3weeks
}
\examples{
air_3wks <- air_3wks(DF)
air_3wks

}
