% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdate.R
\name{bdate}
\alias{bdate}
\title{The base date
for Cows and Heifers}
\usage{
bdate(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
date the base date
}
\description{
The base date
for Cows and Heifers
}
\examples{
bdate <- bdate(DF)
bdate

}
