% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_3wks.R
\name{cr_3wks}
\alias{cr_3wks}
\title{Conception Rate in the last 3 weeks}
\usage{
cr_3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
double conception rate (\%)
}
\description{
(heads_of_pregnancy /
(the_sum_of_AI -  the_sum_of_the_unknown_pregnant_AI))*100 \%
...in_the_last_3weeks
}
\examples{
cr_3wks <- cr_3wks(DF)
cr_3wks

}
