% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmk_DF_RPD.R
\name{frmk_DF_RPD}
\alias{frmk_DF_RPD}
\title{Makes data frame from csvfile}
\usage{
frmk_DF_RPD(csvfile)
}
\arguments{
\item{csvfile}{A name of the csvfile

return A data frame}
}
\description{
Makes data frame from csvfile.
}
\examples{
DF <- frmk_DF_RPD("MH_20220313Cow00.csv")
str(DF)
head(DF)
View(DF)

}
