% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsofp_lst3wks.R
\name{hsofp_lst3wks}
\alias{hsofp_lst3wks}
\title{Heads of pregnancy in the last 3 weeks
for Cows and Heifers}
\usage{
hsofp_lst3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int heads of pregnancy
}
\description{
Heads of pregnancy in the last 3 weeks
for Cows and Heifers
}
\examples{
 hsofp_lst3wks <- hsofp_lst3wks(DF)
 hsofp_lst3wks

}
