% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmr.R
\name{jmr}
\alias{jmr}
\title{JMR}
\usage{
jmr(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period}
}
\value{
double JMR
}
\description{
JMR
}
\examples{
jmr <- jmr(DF,50)
jmr

}
