% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmrH.R
\name{jmrH}
\alias{jmrH}
\title{JMR
for Heifer}
\usage{
jmrH(DFH, VWP)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period (months)}
}
\value{
double JMR
}
\description{
JMR
for Heifer
}
\examples{
jmrH <- jmrH(DFH,13)
jmrH

}
