% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof_unkpr_ai.R
\name{nof_unkpr_ai}
\alias{nof_unkpr_ai}
\title{Count the number of unknown pregnancy AI}
\usage{
nof_unkpr_ai(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
integer
}
\description{
This return is the number of the nonpregnant AI
}
\examples{
nof_unkpr_ai <- nof_unkpr_ai(DF)
nof_unkpr_ai
}
