% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_3wks.R
\name{pr_3wks}
\alias{pr_3wks}
\title{The pregnancy rate in the last 3 weeks}
\usage{
pr_3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
double pregnancy rate in the 3 weeks(\%)
}
\description{
((the_AI_rate_in_the_lst3wks / 100) *
(the_coception_rate_in_the_lst3wks / 100)) * 100(\%)
}
\examples{
 pr_3wks <- pr_3wks(DF)
 pr_3wks

}
