% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotient_21.R
\name{quotient_21}
\alias{quotient_21}
\title{The_days_of_opencow - VWP devide by 21}
\usage{
quotient_21(DF, VWP, i)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}

\item{i}{rowNo}
}
\value{
quotient_21
}
\description{
The_days_of_opencow - VWP devide by 21
}
\examples{
quotient_10 <- quotient_21(DF, 50, 10 )
quotient_10
print("Group 1")
quotient_40 <- quotient_21(DF, 50, 40 )
quotient_40
print("Group 4")
quotient_80 <- quotient_21(DF, 50, 80 )
quotient_80
print("Group 5")

}
