% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotient_21H.R
\name{quotient_21H}
\alias{quotient_21H}
\title{The_days_of_opencow - VWP devide by 21 for Heifer}
\usage{
quotient_21H(DFH, VWP, i)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period (months)}

\item{i}{rowNo}
}
\value{
int quotient_21H
}
\description{
The_days_of_opencow - VWP devide by 21 for Heifer
}
\examples{
quotient_5 <- quotient_21H(DFH, 13, 5 )
quotient_5
print("Group 1")
quotient_20 <- quotient_21H(DFH, 13, 20 )
quotient_20

print("Group 4")
quotient_30 <- quotient_21H(DFH, 13, 30 )
quotient_30
print("Group 5")

}
