% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remainder_21.R
\name{remainder_21}
\alias{remainder_21}
\title{Get the remainder of which The_days_of_opencow - VWP devide by 21}
\usage{
remainder_21(DF, VWP, i)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}

\item{i}{rowNo}
}
\value{
remainder_21
}
\description{
Get the remainder of which The_days_of_opencow - VWP devide by 21
}
\examples{
remainder_10 <- remainder_21(DF, 50, 10 )
remainder_10
print("Group 1")
remainder_40 <- remainder_21(DF, 50, 40 )
remainder_40
print("Group 4")
remainder_80 <- remainder_21(DF, 50, 80 )
remainder_80
print("Group 5")

}
