% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remainder_21H.R
\name{remainder_21H}
\alias{remainder_21H}
\title{Get the remainder of which The_days_of_openheifer - VWP devide by 21}
\usage{
remainder_21H(DFH, VWP, i)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period (months)}

\item{i}{rowNo}
}
\value{
int remainder_21H
}
\description{
Get the remainder of which The_days_of_openheifer - VWP devide by 21
}
\examples{
remainder_5 <- remainder_21H(DFH, 13, 5 )
remainder_5
print("Group 1")
remainder_20 <- remainder_21H(DFH, 13, 20 )
remainder_20
print("Group 4")
remainder_30 <- remainder_21H(DFH, 13, 30 )
remainder_30
print("Group 5")

}
