% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snof_unkpr_cycles.R
\name{snof_unkpr_cycles}
\alias{snof_unkpr_cycles}
\title{Count the number of the unknown pregnancy cycle}
\usage{
snof_unkpr_cycles(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
the total number of the unknown pregnant cows' cycles
}
\description{
This return is the number of the unknown pregnancy cycle
}
\examples{
snof_unkpr_cycles <- snof_unkpr_cycles(DF)
snof_unkpr_cycles
}
