% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snof_unkpr_cyclesH.R
\name{snof_unkpr_cyclesH}
\alias{snof_unkpr_cyclesH}
\title{Count the number of the unknown pregnancy cycle for Heifer}
\usage{
snof_unkpr_cyclesH(DFH)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}
}
\value{
int the total number of the unknown pregnant heifers' cycles
}
\description{
This return is the number of the unknown pregnancy cycle
}
\examples{
snof_unkpr_cyclesH <- snof_unkpr_cyclesH(DFH)
snof_unkpr_cyclesH
}
