% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snofunkpr_cycles.R
\name{snofunkpr_cycles}
\alias{snofunkpr_cycles}
\title{Count the number of the unknown pregnancy cycle}
\usage{
snofunkpr_cycles(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}
}
\value{
the total number of the unknown pregnant cows' cycles
}
\description{
This return is the total of unknown pregnant Cows' cycles
}
\examples{
snofunkpr_cycles <- snofunkpr_cycles(DF, 50)
snofunkpr_cycles
}
