% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snofunkpr_cyclesH.R
\name{snofunkpr_cyclesH}
\alias{snofunkpr_cyclesH}
\title{Count the number of the unknown pregnancy cycle for Heifer
for Heifer
This return is the total of unknown pregnant Heifers' cycles}
\usage{
snofunkpr_cyclesH(DFH, VWP)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period (months)}
}
\value{
the total number of the unknown pregnant heifers' cycles
}
\description{
Count the number of the unknown pregnancy cycle for Heifer
for Heifer
This return is the total of unknown pregnant Heifers' cycles
}
\examples{
snofunkpr_cyclesH <- snofunkpr_cyclesH(DFH, 13)
snofunkpr_cyclesH
}
