% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofai_lst3wks.R
\name{sofai_lst3wks}
\alias{sofai_lst3wks}
\title{Sum of AItimes in the last 3 weeks}
\usage{
sofai_lst3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int AItimes
}
\description{
Sum of AItimes in the last 3 weeks
}
\examples{
sofai_lst3wks <- sofai_lst3wks(DF)
sofai_lst3wks

}
