% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofcycles.R
\name{sofcycles}
\alias{sofcycles}
\title{Get the sum of cycles}
\usage{
sofcycles(DF, VWP)
}
\arguments{
\item{DF}{A Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}
}
\value{
the sum of cycles
}
\description{
Get the sum of cycles
}
\examples{

sofcycles <- sofcycles(DF, 50)
sofcycles

}
