% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofcyclesH.R
\name{sofcyclesH}
\alias{sofcyclesH}
\title{Get the sum of cycles for Heifer}
\usage{
sofcyclesH(DFH, VWP)
}
\arguments{
\item{DFH}{A Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period (months)}
}
\value{
the sum of cycles
}
\description{
Get the sum of cycles for Heifer
}
\examples{

sofcyclesH <- sofcyclesH(DFH, 13)
sofcyclesH

}
