% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofcycls_3wks.R
\name{sofcycls_3wks}
\alias{sofcycls_3wks}
\title{The sum of cycles in the last 3weeks}
\usage{
sofcycls_3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int the sum of cycles in the last 3weeks
}
\description{
The sum of cycles in the last 3weeks
}
\examples{
sofcycls_3wks <- sofcycls_3wks(DF)
sofcycls_3wks

}
