% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofunkp_ai_lst3wks.R
\name{sofunkp_ai_lst3wks}
\alias{sofunkp_ai_lst3wks}
\title{Sum of the unknown pregnant AI in the last 3 weeks}
\usage{
sofunkp_ai_lst3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int AItimes in the last 3 weeks
}
\description{
Sum of the unknown pregnant AI in the last 3 weeks
}
\examples{
sofunkp_ai_lst3wks <- sofunkp_ai_lst3wks(DF)
sofunkp_ai_lst3wks

}
