% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofunkpcycls_3wks.R
\name{sofunkpcycls_3wks}
\alias{sofunkpcycls_3wks}
\title{The sum of unknown pregnant cycles in the last 3 weeks}
\usage{
sofunkpcycls_3wks(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int the sum of unknown pregnant cycles in the last 3 weeks
}
\description{
The sum of unknown pregnant cycles in the last 3 weeks
}
\examples{
sofunkpcycls_3wks <- sofunkpcycls_3wks(DF)
sofunkpcycls_3wks

}
