% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttl_air.R
\name{ttl_air}
\alias{ttl_air}
\title{Total AI rate
for Cow
omit the unknown pregnancy cycles}
\usage{
ttl_air(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}
}
\value{
double total AI rate (\%)
}
\description{
(the_number_of_AIs -  the_number_of_nonpregnant_AIs /
(the_number_of_cycles -
the_number_of_the_unknown_pregnancy_cycles ))*100 \%
}
\examples{
ttl_AIR <- ttl_air(DF, 50)
ttl_AIR

}
