% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttl_cr.R
\name{ttl_cr}
\alias{ttl_cr}
\title{Total conception rate
for Cow and Heifer
(the_number_of_pregnant_cows /
(the_number_of_AIs -  the_number_of_nonpregnant_AIs))*100 \%}
\usage{
ttl_cr(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
double total conception rate (\%)
}
\description{
Total conception rate
for Cow and Heifer
(the_number_of_pregnant_cows /
(the_number_of_AIs -  the_number_of_nonpregnant_AIs))*100 \%
}
\examples{
ttl_CR <- ttl_cr(DF)
ttl_CR

}
