% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttl_pr.R
\name{ttl_pr}
\alias{ttl_pr}
\title{Total pregnancy rate}
\usage{
ttl_pr(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period}
}
\value{
double total pregnancy rate(\%)
}
\description{
omit the unknown pregnant Cows' AI and cycle
}
\details{
((the_total_AI_rate / 100) * (the_total_coception_rate / 100)) * 100(\%)
}
\examples{
 ttlpr <- ttlpr(DF, 50)
 ttlpr

}
