% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttlair.R
\name{ttlair}
\alias{ttlair}
\title{Total AI rate}
\usage{
ttlair(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period}
}
\value{
double total AI rate (\%)
}
\description{
omit the number of the unknown pregnant Cow's AIs
}
\details{
(the_number_of_AIs -  the_number_of_unknown_pregnantCows'_AIs /
(the_number_of_cycles -
the_number_of_the_unknown_pregnant_Cow's_cycles ))*100 \%
}
\examples{
ttlAIR <- ttlair(DF, 50)
ttlAIR

}
