% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttlairH.R
\name{ttlairH}
\alias{ttlairH}
\title{Total AI rate
for Heifer
omit the number of the unknown pregnant Heifer's AIs}
\usage{
ttlairH(DFH, VWP)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{Voluntary Waiting Period (months)}
}
\value{
double total AI rate (\%)
}
\description{
(the_number_of_AIs -  the_number_of_unknown_pregnantCows'_AIs /
(the_number_of_cycles -
the_number_of_the_unknown_pregnant_Heifer's_cycles ))*100 \%
}
\examples{
ttlAIRH <- ttlairH(DFH, 13)
ttlAIRH

}
