% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttlcr.R
\name{ttlcr}
\alias{ttlcr}
\title{Total conception rate}
\usage{
ttlcr(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
double total conception rate (\%)
}
\description{
(the_number_of_pregnant_cows /
(the_number_of_AIs -  the_number_of_unknown_pregnant_AIs))*100 \%
}
\examples{
ttlCR <- ttlcr(DF)
ttlCR

}
