% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttlpr.R
\name{ttlpr}
\alias{ttlpr}
\title{Total pregnancy rate}
\usage{
ttlpr(DF, VWP)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period}
}
\value{
double total pregnancy rate(\%)
}
\description{
omit the unknown pregnant Cows
}
\details{
((the_total_AI_rate / 100) * (the_total_coception_rate / 100)) * 100(\%)
}
\examples{
 ttlpr <- ttlpr(DF, 50)
 ttlpr

}
