% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttlprH.R
\name{ttlprH}
\alias{ttlprH}
\title{Total pregnancy rate
for Heifer
omit the unknown pregnant Heifers}
\usage{
ttlprH(DFH, VWP)
}
\arguments{
\item{DFH}{Dataframe of the farms' reproduction data}

\item{VWP}{int Voluntary Waiting Period (months)}
}
\value{
double total pregnancy rate(\%)
}
\description{
(the_number_of_AIs -  the_number_of_unknown_pregnantCows'_AIs /
(the_number_of_cycles -
the_number_of_the_unknown_pregnant_Cow's_cycles ))*
(the_number_of_pregnant_cows /
(the_number_of_AIs -  the_number_of_unknown_pregnantCows'_AIs))*100 \%
}
\details{
= (the_number_of_pregnant_cows) / (the_number_of_cycles -
the_number_of_the_unknown_pregnant_Cow's_cycles ) * 100 \%

= ((the_total_AI_rate / 100) * (the_total_coception_rate / 100)) * 100(\%)
}
\examples{
 ttlprH <- ttlprH(DFH, 13)
 ttlprH

}
