% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsofpreg.R
\name{hsofpreg}
\alias{hsofpreg}
\title{Count the heads of pregnancy}
\usage{
hsofpreg(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
integer
}
\description{
Count the heads of pregnancy
}
\examples{
DF
hsofpreg <- hsofpreg(DF)
hsofpreg
hsofpregH <- hsofpreg(DFH)
hsofpregH
}
