% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nofhd.R
\name{nofhd}
\alias{nofhd}
\title{The number of head
for Cow and Heifer
DF needs to have only breeding cows or heifers(Group 1 _ 6)}
\usage{
nofhd(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
int The number of head
}
\description{
The number of head
for Cow and Heifer
DF needs to have only breeding cows or heifers(Group 1 _ 6)
}
\examples{
nofhd <- nofhd(DF)
nofhd

}
