% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nofunkpr_ai.R
\name{nofunkpr_ai}
\alias{nofunkpr_ai}
\title{Count the number of the unknown pregnancy AI
for Cow and Heifer
This return is the total of unknown pregnant Cows' AItimes}
\usage{
nofunkpr_ai(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
integer
}
\description{
Count the number of the unknown pregnancy AI
for Cow and Heifer
This return is the total of unknown pregnant Cows' AItimes
}
\examples{
nofunkpr_ai <- nofunkpr_ai(DF)
nofunkpr_ai
}
