% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofaitimes.R
\name{sofaitimes}
\alias{sofaitimes}
\title{Sum up the number of column"AItimes"}
\usage{
sofaitimes(DF)
}
\arguments{
\item{DF}{Dataframe of the farms' reproduction data}
}
\value{
integer
}
\description{
Sum up the number of column"AItimes"
}
\examples{
sofaitimes <- sofaitimes(DF)
sofaitimes
}
